refactor(tangle-dapp): Make account page the homepage
