--- apps/tangle-dapp/app/account/types.ts
@@ -1,3 +1,0 @@
-export type AccountSummaryCardProps = {
-  className?: string;
-};
